requirements.txt
alembic==1.5.4
attrs==20.3.0
autopep8==1.5.5
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
certifi==2020.12.5
chardet==4.0.0
click==8.1.6
coverage==7.2.7
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==2.6.0
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
idna==2.10
importlib-metadata==6.8.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.4
MarkupSafe==2.1.3
packaging==20.9
pluggy==0.13.1
psycopg2==2.9.6
psycopg2-binary==2.9.5
py==1.11.0
pycodestyle==2.6.0
pyparsing==2.4.7
pytest==7.1.1
pytest-cov==2.12.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
tomli==2.0.1
urllib3==1.26.5
Werkzeug==2.3.6
zipp==3.16.2
